pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "utils-api"
version = "0.1.0"
description = "Shared utilities for API services (logging, configuration, helpers)"
authors = [{ name = "Team" }]
requires-python = ">=3.9"
dependencies = [
  "pydantic-settings>=2.1.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["."]
